🧱 Storyline in Place
- All storyline is put into the levels and all phases are present
- Admin does spawn animation when spawning packets
- Packet bombs now explode on Tilemap floors
_ Added a classification key for storyline (for more org)

+++ scenes/scripts/packet_bomb.cs
@@ -29,7 +29,7 @@ public partial class packet_bomb : CharacterBody2D
 
 	private void _on_area_2d_body_entered(Node body)
 	{
-		if (body is StaticBody2D floor) 
+		if (body is StaticBody2D || (body is TileMap)) 
 		{
 			animator.Play("explode");
 		}
@@ -41,7 +41,10 @@ public partial class packet_bomb : CharacterBody2D
 		{
 			if (character is Percy player)
 			{
-				player.kill();	
+				if (animator.IsPlaying() == true)
+				{
+					player.kill();	
+				}
 			}
 		}
 	}
